+++ packages/react-mui-store-file-viewer/src/components/CreateNewShareButton/CreateSharesViaSmsButton/index.tsx
@@ -100,7 +100,7 @@ const CreateSharesViaSmsButton: FC<Props> = ({
         const links = response?.result?.share_link_objects ?? [];
         if (links.length && apiRef.share?.send) {
           const body: ObjectStore.ShareSendRequest = {
-            from_prefix: "",
+            sender_email: "",
             links: [],
           };