+++ packages/react-mui-shared/CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - SelectField
 - PopoutCard component
 
-### **Fixed**
+### Fixed
 
 - useInternalState not triggering callback function