+++ packages/react-mui-store-file-viewer/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.17] - 2024-01-16
+
+### Fixed
+
+- Added loading indicator to download search control
+- Removed uppercase on folder path
+
 ## [0.0.14] - 2024-01-16
 
 ### Added